Add vibrationMotor class for unit testing
This commit adds the vibrationMotor class to the unit tests. The vibrationMotor class is responsible for controlling the vibration motor and includes methods to vibrate the motor with different patterns. This class will be used to test the functionality of the vibration motor and ensure that the vibration patterns are working as expected.